% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkbaseball_functions.R
\name{assess_models}
\alias{assess_models}
\title{A Model Assessment Function}
\usage{
assess_models(final_list, typ)
}
\arguments{
\item{final_list}{large list from \code{eval_in()}.}

\item{typ}{evaluate for all pitches "all" or for a specific count selected from \code{countfull} if \code{varobs}="ct" or from \code{counthalf} if \code{varobs}="ct_red" from original models.}
}
\description{
This function evaluates the models with respect to their average performance across all pitchers. Returns df, \emph{not} large list as other previous functions in list situations.
}
\keyword{assess}
