% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkbaseball_functions.R
\name{big_model}
\alias{big_model}
\title{A Model Maker}
\usage{
big_model(prepped_list)
}
\arguments{
\item{prepped_list}{large list prepared using \code{prep_model_list()}.}
}
\description{
This function creates the HMM for each pitcher given the variable pairs indicated from prep_model_list.
}
\keyword{model}
