% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkbaseball_functions.R
\name{counter_indicesg}
\alias{counter_indicesg}
\title{An Emission Matrix Preformatter}
\usage{
counter_indicesg(dfclean, varhide, varobs, stats, symbs)
}
\arguments{
\item{dfclean}{dataframe cleaned using \code{data_clean()}, can be found in \verb{[[2]]} of the large list.}

\item{varhide}{string name for hidden variable, one of "pitch", "pitch_red1", "pitch_red2".}

\item{varobs}{string name for observed variable, one of "ct" or "ct_red".}

\item{stats}{vector pitch arsenal from \code{fullpitch()}.}

\item{symbs}{vector count list from either \code{countfull} corresponding to \code{varobs}="ct" or \code{counthalf} corresponding to \code{varobs}="ct_red".}
}
\description{
This function preformats the data for an emission matrix to be used in the HMM. Unused on its own unless for debugging, because it always needs debugging.
}
\keyword{emission}
