% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkbaseball_functions.R
\name{data_clean_off}
\alias{data_clean_off}
\title{A Data Cleaning Function FOR OFF SIDES}
\usage{
data_clean_off(file)
}
\arguments{
\item{file}{filename from your wokring directory, ex '/data/wheeler.csv'.}
}
\description{
This function cleans files downloaded directly from a statcast search on baseballsavant.mlb.com. Example search: https://baseballsavant.mlb.com/statcast_search?hfPT=&hfAB=&hfGT=R\%7C&hfPR=&hfZ=&hfStadium=&hfBBL=&hfNewZones=&hfPull=&hfC=&hfSea=2025\%7C&hfSit=&player_type=pitcher&hfOuts=&home_road=&pitcher_throws=&batter_stands=&hfSA=&hfEventOuts=&hfEventRuns=&game_date_gt=&game_date_lt=&hfMo=&hfTeam=&hfOpponent=&hfRO=&position=&hfInfield=&hfOutfield=&hfInn=&hfBBT=&hfFlag=&metric_1=&group_by=name&min_pitches=2000&min_results=25&min_pas=0&sort_col=xwoba&player_event_sort=api_p_release_speed&sort_order=asc&chk_stats_abs=on&chk_stats_hits=on&chk_stats_xwoba=on#results where for each player, click "Graphs" (far right) and then click "Download as CSV". Splits sequences by at-bat.
}
\keyword{clean}
