% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkbaseball_functions.R
\name{dice_sorensen}
\alias{dice_sorensen}
\title{Dice-Sorensen Coefficient Finder}
\usage{
dice_sorensen(x, y)
}
\arguments{
\item{x}{vector of same length as \code{y}.}

\item{y}{vector of same length as \code{x}.}
}
\description{
This function finds the Dice-So/rensen coefficient between two vectors of equal length. This doesn't generalize to other d-s options because I don't need it to.
}
\keyword{dicesorensen}
