% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkbaseball_functions.R
\name{emissionprob}
\alias{emissionprob}
\title{An Emission Probability Matrixer}
\usage{
emissionprob(ctindex, varhide, varobs)
}
\arguments{
\item{ctindex}{preformatted emission data from \code{counter_indicesg()}.}

\item{varhide}{string name for hidden variable, one of "pitch", "pitch_red1", "pitch_red2".}

\item{varobs}{string name for observed variable, one of "ct" or "ct_red".}
}
\description{
This function produces an emission matrix containing the emission probabilities for the counts from each pitch (probability of each pitch-count pair) to be used in the HMM.
}
\keyword{emission}
