% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkbaseball_functions.R
\name{eval_in}
\alias{eval_in}
\title{A Model Evaluator}
\usage{
eval_in(full_list)
}
\arguments{
\item{full_list}{large list from the output of \code{big_model()}.}
}
\description{
This function evaluates each model with respect to all pitches and either 0-0/0-2/3-0/3-2 counts or 0-0/even/hitter/pitcher counts depending on the observed variable specified.
}
\keyword{evaluate}
