% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkbaseball_functions.R
\name{evaluate_viterbi}
\alias{evaluate_viterbi}
\title{A Viterbi Evaluator}
\usage{
evaluate_viterbi(vpred_list, varhide, varobs, typet)
}
\arguments{
\item{vpred_list}{list of Viterbi predictions from \code{season_viterbi()}.}

\item{varhide}{string name for hidden variable, one of "pitch", "pitch_red1", "pitch_red2".}

\item{varobs}{string name for observed variable, one of "ct" or "ct_red".}

\item{typet}{evaluate for all pitches "all" or for a specific count selected from \code{countfull} if \code{varobs}="ct" or from \code{counthalf} if \code{varobs}="ct_red".}
}
\description{
This function evaluates the Viterbi estimates from the HMM using Hamming Distance, Dice-Sorensen, and Smith-Waterman methods, returning a dataframe with estimates for each at-bat.
}
\keyword{viterbi}
