% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkbaseball_functions.R
\name{firstpitches}
\alias{firstpitches}
\title{A First Pitch Probability Finder}
\usage{
firstpitches(df, stats, pitchchoice)
}
\arguments{
\item{df}{dataframe cleaned using \code{data_clean()}, can be found in \verb{[[2]]} of the large list.}

\item{stats}{vector pitch arsenal from \code{fullpitch()}.}

\item{pitchchoice}{string indicating whether to use all unique pitches "pitch", pitches grouped "pitch_red1", or fastball/offspeed only "pitch_red2". Must match pitchchoice used in \code{fullpitch()} for \code{stats}.}
}
\description{
This function finds the probabilities for each pitch type being thrown on the first pitch of any given at-bat by a particular pitcher and returns a dataframe with pitch frequencies to be used in the HMM.
}
\keyword{first}
