% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkbaseball_functions.R
\name{fullpitch}
\alias{fullpitch}
\title{A Pitch Arsenal Finder}
\usage{
fullpitch(df, pitchchoice)
}
\arguments{
\item{df}{dataframe cleaned using \code{data_clean()}, can be found in \verb{[[2]]} of the large list.}

\item{pitchchoice}{string indicating whether to use all unique pitches "pitch", pitches grouped "pitch_red1", or fastball/offspeed only "pitch_red2".}
}
\description{
This function identifies the pitch arsenal used by any specific pitcher and returns a vector of alphabetized pitch types with standard abbreviations.
}
\keyword{arsenal}
