% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkbaseball_functions.R
\name{get_listg}
\alias{get_listg}
\title{A Training Data Setter Lister}
\usage{
get_listg(lst_train, varb)
}
\arguments{
\item{lst_train}{list of training data made using \code{train_test_split()}, can be found in \verb{[[3]]} of the large list.}

\item{varb}{hidden variable to be used in the HMM, in this case "pitch", "pitch_red1", or "pitch_red2".}
}
\description{
This function takes the training data from dataframe form into lists of at-bats with only the necessary information.
}
\keyword{train}
