% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkbaseball_functions.R
\name{get_tprobg}
\alias{get_tprobg}
\title{A Transition Matrixer}
\usage{
get_tprobg(lst_got)
}
\arguments{
\item{lst_got}{training list obtained from \code{get_listg()}.}
}
\description{
This function produces a transition matrix between variables given a list of at-bats to be used in the HMM.
}
\keyword{transition}
