% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkbaseball_functions.R
\name{prep_model_list}
\alias{prep_model_list}
\title{A Data Prepper}
\usage{
prep_model_list(read_list, varhide, varobs)
}
\arguments{
\item{read_list}{large list of data read using \code{read_data_in()}.}

\item{varhide}{string name for hidden variable, one of "pitch", "pitch_red1", "pitch_red2".}

\item{varobs}{string name for observed variable, one of "ct" or "ct_red".}
}
\description{
This function prepares data to be put into a HMM and later evaluated, remains in one large nested list.
}
\keyword{prep}
