% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkbaseball_functions.R
\name{read_data_in_off}
\alias{read_data_in_off}
\title{A Data Reader FOR OFF SIDES}
\usage{
read_data_in_off(names_list)
}
\arguments{
\item{names_list}{list of filepath names.}
}
\description{
This function reads data in from a list of filenames, can just be copied by going to finder and selecting the lot of them. From here on out, evaluation of the multiple pitchers is done in clumps and all data, evaluations, models, and results will exist in a large nested list specific to the pairing of hidden/observed variables.
}
\keyword{read}
