% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkbaseball_functions.R
\name{season_viterbi}
\alias{season_viterbi}
\title{A Viterbifier}
\usage{
season_viterbi(hmm, testing_list, varhide, varobs)
}
\arguments{
\item{hmm}{HMM (list of size 5).}

\item{testing_list}{lst_train list of training data made using \code{train_test_split()}, can be found in \verb{[[4]]} of the large list.}

\item{varhide}{string name for hidden variable, one of "pitch", "pitch_red1", "pitch_red2".}

\item{varobs}{string name for observed variable, one of "ct" or "ct_red".}
}
\description{
This function uses the Viterbi method to produce the HMM estimates for an entire player's training set of data.
}
\keyword{viterbi}
