% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkbaseball_functions.R
\name{split_shuffle}
\alias{split_shuffle}
\title{A Shuffler}
\usage{
split_shuffle(df)
}
\arguments{
\item{df}{dataframe cleaned using \code{data_clean()}.}
}
\description{
This function takes in a dataframe cleaned using \code{data_clean()} and simply shuffles the orders of the at-bats.
}
\keyword{shuffle}
