% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkbaseball_functions.R
\name{test_df}
\alias{test_df}
\title{A Testing Setter}
\usage{
test_df(shuffs)
}
\arguments{
\item{shuffs}{large list shuffled using \code{split_shuffle()}.}
}
\description{
This function takes in a shuffled large list from \code{split_shuffle()} and takes 30\% of the small lists into a testing dataset.
}
\keyword{test}
