% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkbaseball_functions.R
\name{train_df}
\alias{train_df}
\title{A Training Setter}
\usage{
train_df(shuffs)
}
\arguments{
\item{shuffs}{large list shuffled using \code{split_shuffle()}.}
}
\description{
This function takes in a shuffled large list from \code{split_shuffle()} and takes 70\% of the small lists into a training dataset.
}
\keyword{train}
