% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkbaseball_functions.R
\name{train_test_split}
\alias{train_test_split}
\title{A Train/Test Splitter}
\usage{
train_test_split(filename)

train_test_split(filename)

train_test_split(filename)
}
\arguments{
\item{filename}{large list shuffled using \code{split_shuffle()}.}
}
\description{
This function takes in a file downloaded directly from a statcast search on baseballsavant.mlb.com, per instructions in \code{data_clean()}, and splits into 70\% training and 30\% testing data. Results in large list containing player name, cleaned df, training list, and testing list.

This function takes in a file downloaded directly from a statcast search on baseballsavant.mlb.com, per instructions in \code{data_clean()}, and splits into 70\% training and 30\% testing data. Results in large list containing player name, cleaned df, training list, and testing list.

This function takes in a file downloaded directly from a statcast search on baseballsavant.mlb.com, per instructions in \code{data_clean()}, and splits into 70\% training and 30\% testing data. Results in large list containing player name, cleaned df, training list, and testing list.
}
\keyword{split}
